Update README.md
🌐 About Me:
Hey there! I'm Ranit Banerjee, a passionate 3rd-year student in Electrical and Electronics Engineering with a keen interest in the intersection of AI and robotics. Proficient in languages like C, C++, Java, and Python, I'm on a mission to leverage my skills to pioneer advancements in the world of technology. Currently, I'm immersed in a robotics project, infusing AI to enhance its capabilities.